requirements.txt
Babel==2.6.0
dj-database-url==0.5.0
Django==2.1.7
django-autocomplete-light==3.3.2
django-crispy-forms==1.7.2
django-watson==1.5.2
gunicorn==19.9.0
psycopg2==2.7.7
python-docx==0.8.10
whitenoise==4.1.2
